backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
bip-utils==2.9.3
bitarray==2.9.2
boto3==1.34.11
botocore==1.34.162
cbor2==5.6.4
cdp-sdk==0.10.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ckzg==2.0.1
coincurve==20.0.0
crcmod==1.7
cryptography==43.0.1
cytoolz==0.12.3
easyocr==1.7.2
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.10.0
hexbytes==1.2.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
jmespath==1.0.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
opencv-python-headless==4.10.0.84
packaging==24.2
parsimonious==0.10.0
Pillow==10.1.0
py-sr25519-bindings==0.2.0
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyNaCl==1.5.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.5.0
shapely==2.0.6
six==1.16.0
sympy==1.13.1
tifffile==2024.9.20
tokenizers==0.20.3
toolz==0.12.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.2
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
web3==7.2.0
websockets==13.0.1
yarl==1.11.1
